Updating all OpenShift V3 templates.
Motivation:
The OpenShift V3 templates have been updated in the fh-target-openshift3
project and these updates should be reflected in the templates in this
project.

Modifications:
From an OpenShift V3 MBaaS target created from FH Studio we have
exported templates using commands like the following:

$ oc label dc fh-messaging name=fh-messaging
$ oc export svc,dc,route --all --as-template=fh-messaging -o json -l
name=fh-messaging > fh-messaging-template.json

The labeling was required as our DeploymentConfiguration object were
missing the label which would result in no match for the selector.

Result:
It is possible to use the individual templates to spin up a new FH MBaaS
on OpenShift V3 using the commands:

$ oc new-app -f fh-messaging-template.json

JIRA:
https://issues.jboss.org/browse/RHMAP-2723