Adding parameter for the MBaaS Router host name.
Motivation:
Currently, the template contains a hard coded host name for the Router
created. This should be a optional parameter which if not specified will
be generated.

Modifications:
Add a parameter named MBAAS_ROUTER_DNS. If this parameter is not
specified one will be generated with the format:
<4 random alfa numeric characters>-mbaas-local.feedhenry.io

Result:
The MBaaS Router DNS name can now be configured:

$ oc new-app -f fh-mbaas-template-all.json -o json -p
MBAAS_ROUTER_DNS=test-mbaas.domain.com.org