Adding API keys for Metrics and Messaging.
Motivation:
RHMAP-3216 added these keys to the templates in fh-target-openshift3 and
these changes should be reflected in fh-openshift-templates.
We will only need to do thing during the time we are refactoring the
templates and fh-target-openshift3 will be updated to then use the
fh-mbaas-template-all.json from this project.

Modifications:
Added the keys to the the services and added properties for them to be
configred.

Result:
You should be able to manually inspect the properties using the
following oc commands:

$ oc new-project testing
$ oc new-app -f fh-mbaas-template-all.json -o json