Updating Docker image versions
Motivation:
The image versions are currently out of date. This should be updated to
match the versions listed here:
https://mojo.redhat.com/docs/DOC-1049770

Modifications:
Updated the image versions.

Result:
It should be possible to run the following command to create a project
in OpenShift v3 an deploy an mbaas (note that you have to manually pull
down the docker images as specified in the above document):
$ oc new-project test
$ oc new-app /mnt/src/fh-openshift-templates/fh-mbaas-template-all.json