Replace mongodb-service DeploymentConfig with Pods
JIRA: https://issues.jboss.org/browse/RHMAP-3814

Motivation:
We needed finer-grained controls over the MongoDB pods, so that we could
use distinct volumes with them.

Modification:
The scalable DeploymentConfig object was removed from the template, and
3 Pod objects were put in its place. These 3 Pods will be behind the
mongodb-service Service.

A fourth Pod was created to initiate the replicaset. This pod isn't part
of the service, and doesn't get restarted once it completes.

Result:
We should now be in a position to assign different PVs to each of the
mongo Pods, by using different PVCs.

The 'downside' is that we don't have the pushbutton ability to scale the
mongo deployment, but we hadn't tested that very well in any case. We
expect that there will be a new object type in k8s in future that will
better be able to handle this case.