Add PersistentVolumeClaims to Mongo Pods in MBaaS
JIRA: https://issues.jboss.org/browse/RHMAP-3701

Motivation:
We need to be able to persist db data in a reliable place.

Modification:
Three new objects are in the template, of the kind
PersistentVolumeClaim. These request a PersistentVolume that has at
least 1Gi of storage advertised.

The mongo Pods (for the service, not the once-off initiator), now have a
volumes section, that references the corresponding PVC. The mongodb
container in the containers section of the Pod also has a volumeMounts
section that mounts that Pod-specified volume at a provided mountPath in
the container.

Result:
Now that we're using PVCs, we need to ensure that there are enough (3)
available PVs available on the cluster, that match the PVC storage
requirements. Otherwise the Pods will wait in a ContainerCreating
state (and maybe timeout after a period).