Adding a description and iconClass to the template
Motivation:
Currently the template lacks a description and should have one to be
more inline with the templates that are provided out of the box with
OpenShift.

Modifications:
Added a description for the template and also updated the iconClass to
be of type nodejs (was the closes match I found).

Result:
Creating/adding the template to OpenShift using:
$ oc create -n openshift -f fh-mbaas-template-all.json

Listing the templates will now display the description added:
$ oc get templates -n openshift | grep fh-mbaas
fh-mbaas      Red Hat Mobile Backend as a Service template

This description will now also be shown using the Web Console as will
the nodejs icon now be shown (instead of the default lightning icon)