Update to latest image versions
JIRA: https://issues.jboss.org/browse/RHMAP-3660

Motivation:
Get the template to be the place where the image versions change, rather
than fh-supercore config.

Modification:
Update to the latest docker image versions of MBaaS components

Result:
Image versions will be up-to-date for when the core is updated to not
override them.