Use DCs instead of bare Pods for MongoDB
JIRA: https://issues.jboss.org/browse/RHMAP-4082

Motivation:
If an Openshift node evicts one of the MongoDB pods, it won't come back
on another node, if there's no ReplicationController to reschedule
it. If all (or, enough) of the nodes go down (even independently when,
say, the first comes back before the second goes down), then there will
either be no MongoDB Pods, or not enough for replication.

Modification:
Upgrade the Pod objects containing the MongoDB replica members to be
DeploymentConfig objects.

Result:
This will give us a ReplicationController which will try to reschedule
the Pods when possible. It will also allow us to take down and bring
back members manually.