Add UDP port to fh-statsd for $fh.stats
JIRA: https://issues.jboss.org/browse/RHMAP-3250

Motivation:
Cloud apps need to be able to connect to fh-statsd over UDP so that
stats data shows up in the studio for the app.

Modification:
Add a UDP port to fh-stats service.

Result:
UDP connections can be made at port 8081 on the service
IP.

Unfortunately the Openshift router doesn't seem to support UDP, so
putting a Route in front of this port doesn't help. See the following
page for info on the protocols supported by the router:

https://docs.openshift.com/enterprise/3.1/architecture/core_concepts/routes.html