Change restartPolicy of mongo-initiator to Never
Motivation:
We changed this to OnFailure as a hack to try to get lucky when we
didn't know how many Pods we'd get at initiation. It never really gave
us anything though, and there are cases where it would keep
restarting. For example sometimes it would fail near the end with some
initialized Pods, and some uninitialized Pods. When it would restart,
it would fail again with an exception saying that it can't reinitiate
members that are already initiated.

We no longer get the original benefit that we wanted, and for any other
cases it seems like it would be better to fail explicitly rather than
restart indefinitely: if it goes into a restart loop, there's a good
chance that the error from failure #1 will not be the same error in
subsequent failures, and then it's not trivial to find the root cause.

Modification:
Change `restartPolicy` for the mongodb-initiator pod in each of the
templates from `OnFailure` to `Never`.

Result:
The mongodb-initiator pod should now fail if anything goes wrong rather
than restart.