Adding FH_STATSD_API_KEY to fh-mbaas templates.
Motivation:
We want to be able to configure an API key for fh-statsd hence we need
to add an environement variable to both fh-mbaas templates.

Modifications:
Added the FH_STATSD_API_KEY to both templates and to the fh-mbaas
containers env array.

Result:
Is should be possible to override added environment variables
FH_STATSD_API_KEY in the templates:

$ oc process -f fh-mbaas-template-1node.json -v FH_STATSD_API_KEY=testing
| grep -2 FH_STATSD_API_KEY

$ oc process -f fh-mbaas-template-3node.json -v FH_STATSD_API_KEY=testing
| grep -2 FH_STATSD_API_KEY

JIRA:
https://issues.jboss.org/browse/RHMAP-5420