Adding FH_LOG_LEVEL env var to MBaaS components.
Motivation:
Currently, the only way to configure the log level is to update the
configuration file in the underlying node projects that make up the
MBaaS components, that is fh-mbaas, fh-messaging, fh-metrics, and
fh-statsd. To make it easier to debug issues we need a way to toggle the
log level by simply updating an environment variable.

Modifications:
Added FH_LOG_LEVEL with a default value of debug. The idea is
that the MBaaS components will check for this variable and use its value
if it has been set.

Result:
Processing one of the templates should allow for overriding the default
log level:
$ oc process -f fh-mbaas-template-1node.json -v FH_DEFAULT_LOG_LEVEL=error
| grep -5 FH_LOG_LEVEL

Creating a new project using the 1node template and updating the log
level:
$ oc new-project testing
$ oc new-app -f fh-mbaas-template-1node.json -p
FH_DEFAULT_LOG_LEVEL=error

Verify that the log level exist for the deployment configs:
$ oc describe dc | grep FH_LOG_LEVEL

Update the log level for a deployment config:
$ oc env dc/fh-mbaas FH_LOG_LEVEL=info

JIRA:
https://issues.jboss.org/browse/RHMAP-5568