Add npm script to output image tags & command
Motivation:
It's a pain to get the details of the images from a template on a
cluster during development or verification, so that they can be pulled
on the Openshift nodes.

Modification:
Added a script that can be run from the root of the templates directory
that will output a list of the images specified in the template, and
also the command that can be used to pull them all with Docker.

Result:
It is now possible to run `npm run get-images` to get all of the images
defined for an MBaaS, rather than the old way of looking into the
template and manually concatenating the two relevant parameters for each
image and constructing the docker pull commands from those.

Currently the path to the single-node MBaaS template is hardcoded, this
can be addressed in future if necessary.