Add a version script for npm
Motivation:
When we change a template parameter automatically (e.g. a component
version) in Jenkins, we want the ability to be able to also bump the
package/template versions consistently also.

Modification:
Add a script, `./scripts/version.js`, which sets the version of in the
template annotation to be whatever the value of the version in the
`package.json` is. Also, set the `version` script in `package.json` to
run this new script and add any modified files to the index (currently
done outside the javascripts for convenience).

Result:
Running `npm version patch` will patch bump the version in the
`package.json` and the two existing templates. It will also commit these
changes and tag with the version.