Add core template to update-param script
Motivation:
The [fh-openshift-templates-component-update](
http://bob.feedhenry.net:8080/job/fh-openshift-templates-component-update/)
job now gets triggered for `fh-supercore`, and soon for other core
components, so it needs to be able to update the correct
template. Additional templates should be added in the same way.

Modification:
Add `rhmap-core-template.json' to the list of templates in `templateMap`
in `scripts/update-param.js`.

Result:
This new template will also be scanned for the relevant parameter, and
updated if it's found.