Add an SCC for RunAsAny, but has SYS_CHROOT
Motivation:
For gitlab-shell in OpenShift, we run sshd in the container. Running as
root is a requirement for sshd, so we tried to use the `anyuid`
SCC. However, that SCC drops the SYS_CHROOT capability, which means that
sshd can't chroot, which it needs to, when new sessions are created.

Modification:
Added a file for an SCC that gives just enough ability to be able to run
sshd for gitlab-shell.

Result:
The new file here, `scc-anyuid-with-chroot.json` can be used by a
cluster admin (such as the system:admin user), to create the new SCC,
`anyuid-with-chroot`.

To create the new SCC in the local fhcap OpenShift dev VM, run:

```
sudo oc create -f ./fh-openshift-templates/scc-anyuid-with-chroot.json
```

Then once a project has been created, add the SCC to the default
serviceaccount for the project, with the following (replace my-project):

```
sudo oadm policy add-scc-to-user anyuid-with-chroot \
  system:serviceaccount:my-project:default
```