Update automation script to include all templates
Motivation:
Some templates were missing from this, so haven't been getting automatic
update PRs.

Modification:
* Added the missing templates
* Sorted the list in the script
* Fixed the name of the mysql* variables in the script
* Ran the script, which updated some formatting in some templates (this
  would otherwise come in the next automated PR, so it makes sense to do
  it here)

Result:
The script should now be up-to-date. It should be updated with any new
or removed, or moved templates as changes occur, until we decide that we
want it to look for all template files (probably requires some thought).