Update to integrate the httpd template into millicore
Motivation:
In order to share file systems the httpd service is now included into the millicore template

Modifications:
Remove all artifacts and references to fh-mobile-sdks from the dockerfile and update the millicore template to include httpd.

Result:
The httpd servcie will be included in the millicore template and will share the same folder mount point as configured in millicore.

JIRA:
https://issues.jboss.org/browse/RHMAP-7696 (parent)