Fix update param script
The httpd template was removed from fh-openshift-templates in
https://github.com/fheng/fh-openshift-templates/pull/506, because of this
the update param script is breaking when trying to access the file.

This means that version bumps in fh-openshift templates are not currently
working properly.

This removes all references to the httpd template so that the script
will not produce an error.