Hardcode uid for fh-scm; mv scm-data from /tmp
JIRA: https://issues.jboss.org/browse/RHMAP-9404

Motivation:
The ranged user that gets assigned by the "restricted" SCC is not enough
to be able to use SSH, since there's no corresponding user in
/etc/passwd.

Modification:
The "default" user is created in the base template, with uid 1001. This
change just specifies in the OpenShift template that we'd like to run as
that user.

We've created a follow-up issue to investigate using `nss_wrapper` so
that we could go back to using the "restricted" SCC ranged user:

https://issues.jboss.org/browse/RHMAP-9405

While we're here, we've also moved the mount path for the scm-data
volume to be `/opt/scm-data`. Maybe that's not the optimal place, but it
seems better than `/tmp`.

Result:
Since the user we specify in the template isn't provided in the range
that the "restricted" SCC provides, fh-scm will use the
"anyuid-with-chroot" SCC. In any case, fh-scm now works again.