RHMAP-8189 Update shared infra params to use ConfigMaps
Currently the way in which values such as passwords are being shared across
templates are by providing them as parameters in a script.

This updates the infra templates to create ConfigMaps of these values that
can then be used by all components created after this.

The templates which are provided with ConfigMaps are:
* MySQL
* MongoDB