RHMAP-8189 Remove unwanted parameter defaults
Currently the parameters `CLUSTER_HOST` and `CLUSTER_DOMAIN` have
default values which are unwanted.

This updates the default values for these to be empty. The script for the
frontend template has also been updated in order to provide these values.

As the parameters are required the template will not be built unless a value
for these parameters is provided.