RHMAP-8189 Replace millicore-config Secret with generated ConfigMap
Currently the `millicore-config` Secret used by the `millicore` template
has static and incorrect values of "GENERATEME" and "GITGENERATEME". These
must be updated to reflect the service keys contained in the ConfigMaps.

This creates a ConfigMap instead of a secret which uses these values. The
`servicekeys.json` is created, admittedly it's very ugly, with the correct
values.