RHMAP-7790 Add mbaas version info to mbaas template
The system dump tool needs a way to access the version of the mbaas
being used.

This adds a ConfigMap to each mbaas template which contains the
version of the mbaas. This can then be read by the system dump tool
by parsing the output of commands such as:

```
oc get configmap fh-mbaas-info -o yaml
```

or

```
oc get configmap fh-mbaas-info -o json
```

depending on the desired format.