Update for mongodb when running in standalone mode
Motivation:
When running mongodb in standalone mode (i.e no replicaset being used) and the current ose replica is scaled to 0 and then back to 1 or in a failure scenario where the pod is replcaced,
the pod doesn't recognize the replicaset and goes into OTHER status (actaully the member is removed as it can't resolve the dns lookup)

Modifications:
Updates to the bash scripts to ensure no error occurs when the pod is replaced (when trying to create existing users).
Updates to the fh-openshift template to ensure the MONGODB_REPLICA_NAME is empty, as weel as remove the mongodb-initiator pod as it's no longer needed, the relevant users are added in line

Result:
When replaced , either from scaling or failure the pod (mongdb) should mot error and remain in PRIMARY status.

JIRA:
https://issues.jboss.org/browse/RHMAP-9425