Update to allow for replicaset name to configurable
Motivation:
The changes needed here will allow us to set the replicaset name for the on-prem core and MBaaS deploys

Modifications:
Include the MONGODB_REPLICA_NAME envar in the params section and update the various templates to use this paramater.

Result:
This will alow the components to work in both on-prem core (no replicaset enabled) and MBaaS (replicaset enabled).

JIRA:
https://issues.jboss.org/browse/RHMAP-9817