RHMAP-10216 Change mount and repo locations for gitlab shell
JIRA: https://issues.jboss.org/browse/RHMAP-10216

Currently the persistent volume containing gitlab-shell repositories
is changing the permissions of the volume directory.

This adds a volume inside the top-level mount volume and performs permission
changes on those instead.

NOTE: This is part of two changes, the other is in the gitlab shell repository.

Verification:
Something like the following.
* Have a core with this configuration running
* Ensure that the permissions of the PV directory are correct
* Ensure projects can be created and updated (in editor etc.)
* Pull the project locally
* Deploy the project