RHMAP-10960 Add Nagios to the 1 node persistent template
*Motivation*

To allow an easy & repeatable way to install a persistent MBaaS rather
than following through manual steps, or using the demo MBaaS in the
studio. While putting together the script (in this PR), I noticed the 1
node persistent template was missing Nagios.

*Modification*

- Added Nagios to the 1 node persistent template.
- Added a script to install the 1 node persistent template (and allow
  configuring different MBaaS templates to install)
- Included Nagios checks as part of the validation in the install script
- Added a new section to the README for using the script and accessing
  Nagios afterwards

*Verificaiton*

Follow through the added documentation in the README to run the
`create-mbaas.sh` script.
Access the Nagios Dashboard after setting up an MBaaS and verify Nagios
is running OK. Given the modification to the 1 node template to add
Nagios, the verificaiton should make sure all Nagios checks are passing
(with the exception of mongodb-2, mongodb-3 & mongodb::replicaset).
Those failing checks will be addressed in a followup issue

thing